converted script to unittest
